+++ src/Abstractions/NexusMods.Abstractions.GameLocators/GamePath.cs
@@ -1,4 +1,7 @@
+using System.Diagnostics;
 using NexusMods.MnemonicDB.Abstractions;
+using NexusMods.MnemonicDB.Abstractions.Attributes;
+using NexusMods.MnemonicDB.Abstractions.ElementComparers;
 using NexusMods.Paths;
 using NexusMods.Paths.Extensions;
 
@@ -129,4 +132,44 @@ public readonly struct GamePath : IPath<GamePath>, IEquatable<GamePath>, ICompar
     /// is allowed to be a temporary id and will be replaced with the actual id the value is transacted
     /// </summary>
     public (EntityId, LocationId, RelativePath) ToGamePathParentTuple(EntityId id) => (id, LocationId, Path);
+}
+
+/// <summary>
+/// Defines a GamePath attribute.
+/// </summary>
+public class GamePathAttribute(string ns, string name) : ScalarAttribute<GamePath, string>(ValueTags.Utf8, ns, name)
+{
+    /// <inheritdoc />
+    protected override string ToLowLevel(GamePath value)
+    {
+        // TODO: make this a reference or something
+        return $"{value.LocationId.Value}|{value.Path}";
+    }
+
+    /// <inheritdoc />
+    protected override GamePath FromLowLevel(string value, ValueTags tags, AttributeResolver resolver)
+    {
+        var parts = value.Split('|');
+        Debug.Assert(parts.Length == 2);
+        return new GamePath(LocationId.From(parts[0]), RelativePath.FromUnsanitizedInput(parts[1]));
+    }
+}
+
+/// <summary>
+/// An attribute that combines an EntityId, LocationId and RelativePath into a single attribute. This is used to represent GamePaths prefixed
+/// with a parent entity so that range queries only return the paths that are children of the parent entity.
+/// </summary>
+public class GamePathParentAttribute(string ns, string name) : TupleAttribute<EntityId, ulong, LocationId, string, RelativePath, string>(ValueTags.Reference, ValueTags.Ascii, ValueTags.Utf8, ns, name) 
+{
+    /// <inheritdoc />
+    protected override (EntityId, LocationId, RelativePath) FromLowLevel((ulong, string, string) value)
+    {
+        return (EntityId.From(value.Item1), LocationId.From(value.Item2), RelativePath.FromUnsanitizedInput(value.Item3));
+    }
+
+    /// <inheritdoc />
+    protected override (ulong, string, string) ToLowLevel((EntityId, LocationId, RelativePath) value)
+    {
+        return (value.Item1.Value, value.Item2.Value, value.Item3);
+    }
 }
