+++ src/Abstractions/NexusMods.Abstractions.GameLocators/GameStore.cs
@@ -1,5 +1,8 @@
 using System.Diagnostics.CodeAnalysis;
 using JetBrains.Annotations;
+using NexusMods.MnemonicDB.Abstractions;
+using NexusMods.MnemonicDB.Abstractions.Attributes;
+using NexusMods.MnemonicDB.Abstractions.ElementComparers;
 using TransparentValueObjects;
 
 namespace NexusMods.Abstractions.GameLocators;
@@ -51,4 +54,22 @@ public readonly partial struct GameStore
     /// Manually added.
     /// </summary>
     public static readonly GameStore ManuallyAdded = From("Manually Added");
+}
+
+/// <summary>
+/// An attribute that contains the name of a game store.
+/// </summary>
+public class GameStoreAttribute(string ns, string name) : ScalarAttribute<GameStore, string>(ValueTags.Ascii, ns, name)
+{
+    /// <inheritdoc />
+    protected override string ToLowLevel(GameStore value)
+    {
+        return value.Value;
+    }
+
+    /// <inheritdoc />
+    protected override GameStore FromLowLevel(string value, ValueTags tag, AttributeResolver resolver)
+    {
+        return GameStore.From(value);
+    }
 }
