+++ src/Abstractions/NexusMods.Abstractions.GameLocators/NexusMods.Abstractions.GameLocators.csproj
@@ -10,6 +10,7 @@
     </ItemGroup>
     <ItemGroup>
       <ProjectReference Include="..\..\Extensions\NexusMods.Extensions.Hashing\NexusMods.Extensions.Hashing.csproj" />
+      <ProjectReference Include="..\NexusMods.Abstractions.NexusWebApi\NexusMods.Abstractions.NexusWebApi.csproj" />
     </ItemGroup>
     <ItemGroup>
       <Reference Include="NexusMods.Abstractions.NexusWebApi">
