+++ src/Abstractions/NexusMods.Abstractions.Loadouts/GameInstallMetadata.cs
@@ -1,4 +1,5 @@
 using NexusMods.Abstractions.DiskState;
+using NexusMods.Abstractions.GameLocators;
 using NexusMods.Abstractions.MnemonicDB.Attributes;
 using NexusMods.Abstractions.NexusWebApi.Types.V2;
 using NexusMods.MnemonicDB.Abstractions.Attributes;
