+++ src/Abstractions/NexusMods.Abstractions.Loadouts/LoadoutItem.cs
@@ -1,6 +1,4 @@
 using JetBrains.Annotations;
-using NexusMods.Abstractions.Library.Models;
-using NexusMods.Abstractions.MnemonicDB.Attributes;
 using NexusMods.MnemonicDB.Abstractions.Attributes;
 using NexusMods.MnemonicDB.Abstractions.Models;
