+++ src/Abstractions/NexusMods.Abstractions.Loadouts/LoadoutItemWithTargetPath.cs
@@ -1,4 +1,5 @@
 using JetBrains.Annotations;
+using NexusMods.Abstractions.GameLocators;
 using NexusMods.Abstractions.MnemonicDB.Attributes;
 using NexusMods.MnemonicDB.Abstractions.Models;
