--- src/Abstractions/NexusMods.Abstractions.MnemonicDB.Attributes/GameDomainAttribute.cs
@@ -1,24 +1,0 @@
-using NexusMods.Abstractions.GameLocators;
-using NexusMods.MnemonicDB.Abstractions;
-using NexusMods.MnemonicDB.Abstractions.Attributes;
-using NexusMods.MnemonicDB.Abstractions.ElementComparers;
-
-namespace NexusMods.Abstractions.MnemonicDB.Attributes;
-
-/// <summary>
-/// MnemonicDB attribute for the GameDomain type.
-/// </summary>
-public class GameDomainAttribute(string ns, string name) : ScalarAttribute<GameDomain, string>(ValueTags.Ascii, ns, name)
-{
-    /// <inheritdoc />
-    protected override string ToLowLevel(GameDomain value)
-    {
-        return value.Value;
-    }
-
-    /// <inheritdoc />
-    protected override GameDomain FromLowLevel(string value, ValueTags tag, AttributeResolver resolver)
-    {
-        return GameDomain.From(value);
-    }
-}
