+++ src/Abstractions/NexusMods.Abstractions.NexusWebApi/IGameDomainToGameIdMappingCache.cs
@@ -1,6 +1,6 @@
 using DynamicData.Kernel;
-using NexusMods.Abstractions.GameLocators;
 using NexusMods.Abstractions.NexusWebApi.Types.V2;
+using GameDomain = NexusMods.Abstractions.NexusWebApi.Types.GameDomain;
 namespace NexusMods.Abstractions.NexusWebApi;
 
 /// <summary>
