+++ src/Networking/NexusMods.Networking.NexusWebApi/V1Interop/GameDomainToGameIdMapping.cs
@@ -1,4 +1,5 @@
 using NexusMods.Abstractions.MnemonicDB.Attributes;
+using NexusMods.Abstractions.NexusWebApi.Types;
 using NexusMods.Abstractions.NexusWebApi.Types.V2;
 using NexusMods.MnemonicDB.Abstractions.Models;
 namespace NexusMods.Networking.NexusWebApi.V1Interop;
