+++ tests/Networking/NexusMods.Networking.NexusWebApi.Tests/GameDomainToGameIdMappingCacheTests.cs
@@ -1,5 +1,6 @@
 using FluentAssertions;
 using NexusMods.Abstractions.GameLocators;
+using NexusMods.Abstractions.NexusWebApi.Types;
 using NexusMods.Abstractions.NexusWebApi.Types.V2;
 using NexusMods.Networking.NexusWebApi.V1Interop;
 namespace NexusMods.Networking.NexusWebApi.Tests;
